llm_service/requirements.txt
fastapi== 0.110.0
uvicorn==0.29.0
requests==2.31.0
langchain==0.1.10
python-dotenv==1.0.1
postgres==4.0
psycopg2-binary==2.9.9
pgvector==0.2.5
pydantic==2.6.3
langchain-community==0.0.25
sentence-transformers==2.2.2
InstructorEmbedding==1.0.1
PyPDF2==3.0.1
langserve==0.1.1
